requirements.txt
Flask==0.10.1
Flask-Assets==0.8
Flask-SQLAlchemy==0.16
Flask-Script==0.5.3
SQLAlchemy==0.8.1
alembic==0.5.0
cssmin==0.1.4
jsmin==2.0.3
python-dateutil==2.1
six==1.3.0
milieu==0.1.3
gunicorn==18.0
redis==2.8.0
Flask-RESTful==0.2.8
python-bcrypt==0.3.2
pygeoip==0.3.0
webassets-recess==0.0.0
twilio==3.6.4
boto==2.23.0
plant==0.1.1
https://codeload.github.com/abourget/gevent-socketio/tar.gz/HEAD
